:pencil: Incorporate specs from notion
